requirements.txt
torch
gym
mss
pydirectinput
selenium
numpy
opencv-python
easyocr
chromedriver_binary
pillow
